design: mainpage 구조 변경
