+++ client/src/components/Footer/Footer.scss
@@ -7,6 +7,8 @@
   justify-content: space-evenly;
   align-items: center;
   border-top: 1px solid #a9a9a9;
+  position: relative;
+  top: 100vh;
 }
 
 .LogoBox {
