+++ client/src/components/Header/Header.scss
@@ -1,9 +1,8 @@
 @import "../../utils";
 
-.Header {
+.header {
   display: flex;
   box-sizing: border-box;
-  box-shadow: 0px 15px 10px -15px rgba(57, 63, 72, 0.3);
   justify-content: space-between;
   align-items: center;
   height: 63px;
@@ -12,44 +11,113 @@
   top: 0;
   left: 0;
   right: 0;
-  background-color: white;
+  background-color: transparent;
   z-index: 10;
+  color: white;
 
   .menu-items {
     display: flex;
     align-items: center;
   }
-  .menu-item,
-  .clock {
+  .menu-item {
     margin: 0 8px;
     cursor: pointer;
     font-size: 1.2rem;
     font-weight: 900;
-    color: black;
+    color: white;
   }
 
   .menu-item:hover {
     color: #faab78;
-    text-shadow: 2px 2px 2px #faab78;
   }
 
-  .logo {
+  .logo,
+  .about-us {
     cursor: pointer;
     font-size: 1.2rem;
     font-weight: 800;
-    color: black;
+    color: white;
+    margin: 0 15px;
 
     &:hover {
-      color: black;
+      color: #faab78;
     }
+  }
+
+  .logo:hover {
+    color: white;
   }
 
   .loginBtn,
   .logoutBtn {
-    border: 1px solid black;
+    border: 1px solid white;
     border-radius: 5px;
     padding: 5px 10px;
-    color: black;
+    color: white;
+    font-weight: 700;
+
+    &:hover {
+      background-color: #faab78;
+      border: 1px solid #faab78;
+      color: white;
+    }
+  }
+}
+
+.header-scroll {
+  display: flex;
+  box-sizing: border-box;
+  justify-content: space-between;
+  align-items: center;
+  height: 63px;
+  padding: 0 30px;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  background-color: black;
+  z-index: 10;
+  color: white;
+
+  .menu-items {
+    display: flex;
+    align-items: center;
+  }
+  .menu-item {
+    margin: 0 8px;
+    cursor: pointer;
+    font-size: 1.2rem;
+    font-weight: 900;
+    color: white;
+  }
+
+  .menu-item:hover {
+    color: #faab78;
+  }
+
+  .logo,
+  .about-us {
+    cursor: pointer;
+    font-size: 1.2rem;
+    font-weight: 800;
+    color: white;
+    margin: 0 15px;
+
+    &:hover {
+      color: #faab78;
+    }
+  }
+
+  .logo:hover {
+    color: white;
+  }
+
+  .loginBtn,
+  .logoutBtn {
+    border: 1px solid white;
+    border-radius: 5px;
+    padding: 5px 10px;
+    color: white;
     font-weight: 700;
 
     &:hover {
