--- client/src/components/MainPage/Main.scss
@@ -1,102 +1,0 @@
-@import "../../utils";
-
-.MainPage {
-  margin-bottom: 250px;
-}
-
-.SlideContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 63px;
-
-  div {
-    transition: all 0.3s ease-out;
-  }
-}
-
-.Slide {
-  display: flex;
-  align-items: center;
-  .window {
-    width: 100vw;
-    height: 750px;
-    overflow: hidden;
-
-    .flexbox {
-      display: flex;
-    }
-  }
-}
-
-.SlideBtn {
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  font-size: 3.3rem;
-  color: white;
-  padding: 0 10px;
-}
-
-.Slideimg {
-  width: 100vw;
-  height: 750px;
-  background-position: 50% 50%;
-  background-size: 100% 750px;
-  background-repeat: no-repeat;
-  flex: none;
-
-  .MoveSlideBtns {
-    display: flex;
-    justify-content: space-between;
-    position: relative;
-    top: 40%;
-    z-index: 10;
-  }
-
-  .move {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    top: 400px;
-
-    a {
-      color: white;
-      font-weight: 700;
-      font-size: 1.3rem;
-      padding: 15px;
-      border-radius: 10px;
-      margin-bottom: 30px;
-    }
-  }
-
-  .CarouselTitle {
-    display: inline-block;
-    position: relative;
-
-    &#first {
-      color: white;
-      text-shadow: 2px 3px 3px black;
-      position: relative;
-      left: 150px;
-      top: 300px;
-      font-size: 2rem;
-      h2 {
-        font-size: 3rem;
-      }
-    }
-
-    &#second {
-      color: white;
-      position: relative;
-      font-size: 2rem;
-      top: 190px;
-      left: 38%;
-      text-align: center;
-      p {
-        font-size: 1.5rem;
-      }
-    }
-  }
-}
