--- client/src/components/MainPage/Search.scss
@@ -1,95 +1,0 @@
-@import "../../utils";
-
-.Search {
-  width: 88vw;
-  margin: auto;
-  margin-top: 100px;
-}
-
-.search_area {
-  display: flex;
-  justify-content: space-between;
-  padding: 0 30px;
-}
-
-.ui.search .prompt {
-  width: 300px;
-  cursor: pointer;
-}
-
-ul.filter_list {
-  display: flex;
-}
-
-ul.filter_list > li {
-  margin: 10px;
-  font-size: 1.5rem;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.Cards {
-  overflow-x: scroll;
-
-  padding: 15px;
-  height: 500px;
-  margin-top: 50px;
-}
-
-.cardBox {
-  display: flex;
-  flex-wrap: wrap;
-  margin-left: 40px;
-}
-
-.place_name {
-  margin-top: 10px;
-}
-
-.imgContent {
-  margin: auto;
-  width: 250px;
-  height: 200px;
-}
-
-.items_img {
-  width: 100%;
-  height: 200px;
-  background-color: aquamarine;
-}
-
-.card {
-  width: 270px;
-  height: 300px;
-  margin: 10px;
-  padding: 5px;
-  border: 2px solid black;
-  border-radius: 5px;
-  cursor: pointer;
-
-  &:hover {
-    transform: translateY(-10px);
-    box-shadow: 5px 5px 10px #a9a9a9;
-  }
-}
-
-.card {
-  margin: 1rem 0.5rem;
-}
-
-#Search {
-  height: 100px;
-  width: 100%;
-}
-
-.active {
-  color: #faab78;
-}
-
-.hidden {
-  display: none;
-}
-
-.cardContent .likeImg {
-  width: 20px;
-}
