+++ client/src/pages/MainPage.js
@@ -1,13 +1,13 @@
 import React from "react";
-import Search from "../components/MainPage/Search";
-import Main from "../components/MainPage/Main";
+import CardSection from "../components/MainPage/CardSection";
+import IntroSection from "../components/MainPage/IntroSection";
 import SideBar from "../components/SideBar/SideBar";
 
 const MainPage = () => {
   return (
     <main className="MainPage">
-      <Main />
-      <Search />
+      <IntroSection />
+      <CardSection />
       <SideBar />
     </main>
   );
